Project 13 intiial changes

+++ controllers/clothingItems.js
@@ -3,6 +3,7 @@ const {
   INVALID_DATA_ERROR,
   NOTFOUND_ERROR,
   DEFAULT_ERROR,
+  FORBIDDEN_ERROR,
 } = require("../utils/error");
 
 const createItem = (req, res) => {
@@ -58,68 +59,40 @@ const updateItem = (req, res) => {
     });
 };
 
-// const deleteItem = (req, res) => {
-//   const { itemId } = req.params;
-//   console.log(req.params);
-//   const { _id: userId } = req.user;
-//   console.log(req.user);
-
-//   ClothingItem.findOne({ _id: itemId });
-//   console.log({ _id: itemId });
-//   console
-//     .log({ owner: userId })
-//     .then(itemId => {
-//       if (!itemId) {
-//         res.status(404).send({ message: "Item not found" });
-//       } else {
-//         ClothingItem.deleteOne({ _id: itemId, owner: userId })
-//           .then(() => {
-//             res.status(200).send({ data: item });
-//           })
-//           .catch((error) => {
-//             res
-//               .status(500)
-//               .send({ message: "An error has occurred on the server" });
-//           });
-//       }
-//     })
-//     .catch((error) => {
-//       if (error.name === "CastError") {
-//         res.status(400).send({ message: "Invalid item ID" });
-//       } else {
-//         res
-//           .status(500)
-//           .send({ message: "An error has occurred on the server" });
-//       }
-//     });
-// };
-
 const deleteItem = (req, res) => {
   const { itemId } = req.params;
   const { _id: userId } = req.user;
 
-  ClothingItem.findOne({ _id: itemId, owner: userId }) // Find item with matching item ID and owner ID
+  ClothingItem.findOne({ _id: itemId })
     .then((item) => {
       if (!item) {
-        res.status(404).send({ message: "Item not found" }); // Item not found, send 404 response
+        res.status(NOTFOUND_ERROR.error).send({ message: "Item not found" });
       } else {
-        ClothingItem.deleteOne({ _id: itemId, owner: userId })
-          .then(() => {
-            res.status(200).send({ data: item }); // Item found and deleted, send 200 response
-          })
-          .catch((error) => {
-            res
-              .status(500)
-              .send({ message: "An error has occurred on the server" });
+        if (item.owner !== userId) {
+          res.status(FORBIDDEN_ERROR.error).send({
+            message: "Unauthorized: Only the card owner can delete it",
           });
+        } else {
+          ClothingItem.deleteOne({ _id: itemId, owner: userId })
+            .then(() => {
+              res.status(200).send({ message: "Item deleted successfully" });
+            })
+            .catch((error) => {
+              res
+                .status(DEFAULT_ERROR.error)
+                .send({ message: "An error has occurred on the server" });
+            });
+        }
       }
     })
     .catch((error) => {
       if (error.name === "CastError") {
-        res.status(400).send({ message: "Invalid item ID" }); // Invalid item ID, send 400 response
+        res
+          .status(INVALID_DATA_ERROR.error)
+          .send({ message: "Invalid item ID" });
       } else {
         res
-          .status(500)
+          .status(DEFAULT_ERROR.error)
           .send({ message: "An error has occurred on the server" });
       }
     });
